// src/app/Models/Employee.ts
export class Employee{
    id: string="";
    firstName: string="";
    lastName: string="";
    email: string="";
    mobileNumber: number=0;
    dateOfBirth: string="";
    dateOfJoin: string="";
    location: string="";
    jobTitle: string="";
    department: string="";
    manager: string="";
    project: string="";
  }